requirements.txt
onnxruntime
pillow
opencv-python
pandas
numpy
shapely
pyclipper
pyyaml
flask
fitz
frontend
PyMuPDF==1.19
openpyxl
premailer
